+++ client/Components/GameBoard/GameBoardLayouts/JoustGameBoardLayout.jsx
@@ -1,16 +1,16 @@
-import React, { useCallback, useMemo } from 'react';
-import PlayerRow from './PlayerRow';
-import PlayerBoard from './PlayerBoard';
-import Droppable from './Droppable';
-import { sendShowDrawDeckMessage, sendShuffleDeckMessage } from '../../redux/reducers/game';
+import React, { useCallback } from 'react';
+import PlayerRow from '../PlayerRow';
+import PlayerBoard from '../PlayerBoard';
+import Droppable from '../Droppable';
+import { sendShowDrawDeckMessage, sendShuffleDeckMessage } from '../../../redux/reducers/game';
 import { useDispatch } from 'react-redux';
-import PlayerStats from './PlayerStats';
+import PlayerStats from '../PlayerStats';
 import classNames from 'classnames';
-import SideBoardPanel from './SideBoardPanel';
+import SideBoardPanel from '../SideBoardPanel';
 
-const MeleeGameBoardLayout = ({
+const JoustGameBoardLayout = ({
     thisPlayer,
-    otherPlayers,
+    otherPlayer,
     onCardClick,
     onMenuItemClick,
     onSettingsClick,
@@ -21,9 +21,8 @@ const MeleeGameBoardLayout = ({
     const dispatch = useDispatch();
 
     const renderPlayerBoard = useCallback(
-        (player, side, hasSidePanel) => {
+        (player, side) => {
             const isMe = thisPlayer && player === thisPlayer;
-            const isActivePrompt = player.isActivePrompt;
 
             const playerBoard = (
                 <PlayerBoard
@@ -37,8 +36,7 @@ const MeleeGameBoardLayout = ({
 
             const wrapperClassName = classNames('flex flex-grow', {
                 'flex-col': side === 'top',
-                'flex-col-reverse': side === 'bottom',
-                'bg-blue-300/5': isActivePrompt
+                'flex-col-reverse': side === 'bottom'
             });
 
             return (
@@ -51,7 +49,6 @@ const MeleeGameBoardLayout = ({
                         onSettingsClick={isMe ? onSettingsClick : undefined}
                         onChatToggle={isMe ? onChatToggle : undefined}
                         unreadMessages={isMe ? unreadMessages : undefined}
-                        seatNo={player.seatNo} // Melee only
                     />
                     <PlayerRow
                         agendas={player.agendas}
@@ -97,7 +94,6 @@ const MeleeGameBoardLayout = ({
                             thisPlayer={thisPlayer}
                             isMe={isMe}
                             side={side}
-                            collapsable={!hasSidePanel}
                         />
                     </div>
                 </div>
@@ -114,55 +110,14 @@ const MeleeGameBoardLayout = ({
             unreadMessages
         ]
     );
-
-    // Since rendering must be in top/bottom groups (with bottom first), we must arrange
-    // in seat order, then rotate boards around to ensure thisPlayer is in the bottom left
-    const players = useMemo(() => {
-        const playersInSeatOrder = [thisPlayer, ...otherPlayers].sort(
-            (a, b) => a.seatNo - b.seatNo
-        );
-        // Rotate so thisPlayer is first index
-        while (playersInSeatOrder[0].seatNo !== thisPlayer.seatNo) {
-            playersInSeatOrder.push(playersInSeatOrder.shift());
-        }
-        const numPlayers = playersInSeatOrder.length;
-        const ordered = [];
-        for (let i = 0; i < numPlayers; i++) {
-            if (i === 0 || i % 2 === 1) {
-                ordered.push(playersInSeatOrder.shift());
-            } else {
-                ordered.push(playersInSeatOrder.pop());
-            }
-        }
-        return ordered;
-    }, [otherPlayers, thisPlayer]);
-
-    const remaining = [...players];
-    const playerBoardGrid = [];
-    const isOddPlayers = players.length % 2 !== 0;
-    for (let column = 0; column < Math.floor(players.length / 2); column++) {
-        const bottom = remaining.shift();
-        // If we have an odd number of players, we ensure thisPlayer's board will span the width of
-        // the first 2 above them, with remaining players rendered normally
-        const top = column === 0 && isOddPlayers ? remaining.splice(0, 2) : remaining.shift();
-
-        playerBoardGrid.push(
-            <div key={column} className='flex flex-col flex-grow'>
-                {Array.isArray(top) ? (
-                    <div className='flex flex-grow'>
-                        {top.map((player, index) =>
-                            renderPlayerBoard(player, 'top', index === 0 && column === 0)
-                        )}
-                    </div>
-                ) : (
-                    renderPlayerBoard(top, 'top', column === 0)
-                )}
-                {renderPlayerBoard(bottom, 'bottom', column === 0)}
+    return (
+        <div className='flex min-h-full'>
+            <div className='flex flex-col flex-grow'>
+                {renderPlayerBoard(otherPlayer, 'top')}
+                {renderPlayerBoard(thisPlayer, 'bottom')}
             </div>
-        );
-    }
-
-    return <div className='flex h-full'>{playerBoardGrid}</div>;
+        </div>
+    );
 };
 
-export default MeleeGameBoardLayout;
+export default JoustGameBoardLayout;
